[NUI] Allows user to set the height of the top and bottom borders respectively. BorderHeight is the default setting. After the initial height is set as the BorderHeight value, the user can set the height of the border as SizeHeight.

+++ src/Tizen.NUI/src/public/Window/IBorderInterface.cs
@@ -41,9 +41,10 @@ namespace Tizen.NUI
 
         /// <summary>
         /// The height of the border.
+        /// This value is the initial value used when creating borders.
         /// </summary>
         [EditorBrowsable(EditorBrowsableState.Never)]
-        public uint BorderHeight {get;}
+        public float BorderHeight {get;}
 
         /// <summary>
         /// The minimum size by which the window will small.
